+++ include/vm/virtual_machine.h
