+++ src/ast/block.cpp
@@ -1,4 +1,4 @@
-#include "block.h"
+#include "ast/block.h"
 
 namespace apus {
